pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-qualitative-analysis"
version = "0.1.0"
description = "Initial skeleton for AI-assisted qualitative analysis workflows."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "litellm>=1.82.0",
  "polars>=1.0.0",
  "pydantic>=2.12.0",
  "PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0.0",
  "ruff>=0.11.0",
]

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]
xfail_strict = true

[tool.ruff]
line-length = 100
src = ["src", "tests"]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
